增加当B站API调用失败时的处理

+++ src/main/java/io/koschicken/bean/bilibili/space/Space.java
@@ -11,6 +11,7 @@ import org.apache.commons.io.FileUtils;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import java.time.LocalDateTime;
 import java.util.Objects;
 
 import static io.koschicken.constants.Constants.COMMON_CONFIG;
@@ -21,6 +22,8 @@ import static org.springframework.util.ResourceUtils.isUrl;
 public class Space {
 
     private static final String LIVE_TEMP_FOLDER = "./temp/bilibili/live/";
+    private static final String USER_AGENT = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36 Edg/108.0.1462.76";
+    private static LocalDateTime next = null;
 
     static {
         File liveFolder = new File(LIVE_TEMP_FOLDER);
@@ -82,29 +85,39 @@ public class Space {
 
     public static Space getSpace(String mid) throws IOException {
         String url = "https://api.bilibili.com/x/space/acc/info?mid=" + mid;
+        if (next != null && LocalDateTime.now().isBefore(next)) {
+            return null;
+        }
         String json = HttpUtils.get(url, COMMON_CONFIG.getBilibiliCookie());
         JSONObject jsonObject = JSON.parseObject(json);
-        String data = jsonObject.getString("data");
-        if (data != null) {
-            Space space = JSON.parseObject(data, Space.class);
-            if (Objects.nonNull(space)) {
-                LiveRoom liveRoom = space.getLiveRoom();
-                if (Objects.nonNull(liveRoom)) {
-                    String coverUrl = liveRoom.getCover().replace("http:", "https:");
-                    boolean isUrl = isUrl(coverUrl);
-                    if (isUrl) {
-                        URL imageUrl = new URL(coverUrl);
-                        String fileName = BilibiliUtils.getImageName(imageUrl);
-                        File cover = new File(LIVE_TEMP_FOLDER + fileName);
-                        FileUtils.deleteQuietly(cover);
-                        FileUtils.touch(cover);
-                        FileUtils.copyURLToFile(imageUrl, cover);
-                        liveRoom.setCoverFile(cover);
-                        space.setLiveRoom(liveRoom);
+        Integer code = jsonObject.getInteger("code");
+        if (code == 0) {
+            String data = jsonObject.getString("data");
+            if (data != null) {
+                Space space = JSON.parseObject(data, Space.class);
+                if (Objects.nonNull(space)) {
+                    LiveRoom liveRoom = space.getLiveRoom();
+                    if (Objects.nonNull(liveRoom)) {
+                        String coverUrl = liveRoom.getCover().replace("http:", "https:");
+                        boolean isUrl = isUrl(coverUrl);
+                        if (isUrl) {
+                            URL imageUrl = new URL(coverUrl);
+                            String fileName = BilibiliUtils.getImageName(imageUrl);
+                            File cover = new File(LIVE_TEMP_FOLDER + fileName);
+                            FileUtils.deleteQuietly(cover);
+                            FileUtils.touch(cover);
+                            FileUtils.copyURLToFile(imageUrl, cover);
+                            liveRoom.setCoverFile(cover);
+                            space.setLiveRoom(liveRoom);
+                        }
                     }
                 }
+                return space;
             }
-            return space;
+        } else {
+            next = LocalDateTime.now().plusMinutes(15);
+            String message = jsonObject.getString("message");
+            log.error("获取用户信息失败，mid={}，message={}", mid, message);
         }
         return null;
     }
